+++ src/Routemeister/MessageRoute.cs
@@ -1,27 +1,32 @@
 ﻿using System;
 using System.Linq;
-using System.Threading.Tasks;
 
 namespace Routemeister
 {
     public class MessageRoute : IEquatable<MessageRoute>
     {
         public Type MessageType { get; }
-        public Func<MessageEnvelope, Task>[] Actions { get; }
+        public IMessageHandlerAction[] Actions { get; }
 
-        internal MessageRoute(Type messageType)
+        public static MessageRoute Empty(Type messageType)
+        {
+            return new MessageRoute(messageType);
+        }
+
+        private MessageRoute(Type messageType)
         {
             if (messageType == null)
                 throw new ArgumentNullException(nameof(messageType));
 
             MessageType = messageType;
-            Actions = new Func<MessageEnvelope, Task>[0];
+            Actions = new IMessageHandlerAction[0];
         }
 
-        public MessageRoute(Type messageType, Func<MessageEnvelope, Task>[] actions) : this(messageType)
+        public MessageRoute(Type messageType, IMessageHandlerAction[] actions) : this(messageType)
         {
             if (actions == null)
                 throw new ArgumentNullException(nameof(actions));
+
             if (!actions.Any())
                 throw new ArgumentException("A message route must have actions. No actions were passed.");