+++ src/Routemeister/Routers/SequentialAsyncMessageRouter.cs
@@ -5,24 +5,29 @@ namespace Routemeister.Routers
 {
     public class SequentialAsyncMessageRouter : IAsyncMessageRouter
     {
-        protected MessageRoutes MessageRoutes { get; }
+        private readonly MessageHandlerCreator _messageHandlerCreator;
+        private readonly MessageRoutes _messageRoutes;
 
-        public SequentialAsyncMessageRouter(MessageRoutes messageRoutes)
+        public SequentialAsyncMessageRouter(MessageHandlerCreator messageHandlerCreator, MessageRoutes messageRoutes)
         {
+            if (messageHandlerCreator == null)
+                throw new ArgumentNullException(nameof(messageHandlerCreator));
+
             if (messageRoutes == null)
                 throw new ArgumentNullException(nameof(messageRoutes));
 
-            MessageRoutes = messageRoutes;
+            _messageHandlerCreator = messageHandlerCreator;
+            _messageRoutes = messageRoutes;
         }
 
         public async Task RouteAsync<T>(T message)
         {
             var messageType = message.GetType();
-            var route = MessageRoutes.GetRoute(messageType);
+            var route = _messageRoutes.GetRoute(messageType);
             var envelope = new MessageEnvelope(message, messageType);
 
             foreach (var action in route.Actions)
-                await action(envelope).ConfigureAwait(false);
+                await action.Invoke(_messageHandlerCreator(action.HandlerType,envelope), envelope.Message).ConfigureAwait(false);
         }
     }
 }