+++ ReleaseNotes.md
@@ -1,5 +1,13 @@
 #Release notes
 
+## v1.0.0 - 2016-03-22
+With this rewrite, the API is now considered stable and no more breaking changes are planned.
+
+- **[Breaking]:** Renamed: `MiddlewareEnabledSequentialAsyncMessageRouter` to `MiddlewareEnabledAsyncMessageRouter`.
+- **[Breaking]:** The `MessageRouteFactory` does not depend on a `MessageHandlerCreator` anymore. The creator is instead injected to any of the pre-made routers.
+- **[Breaking]:** The routers now accept a `MessageHandlerCreator`. This is where you inject your resolving IoC-container.
+- **[Breaking]:** A `MessageRoute` now expose `IMessageHandlerAction[]` instead of `Func<MessageEnvelope, Task>`
+
 ## v0.3.0 - 2016-03-18
 - **[Breaking]:** The resolver that you pass to the `MessageRouteFactory` is now a delegate `MessageHandlerCreator` with a signature `object MessageHandlerCreator(Type messageHandlerContainerType, MessageEnvelope envelope)`. More info in the README.md
 - **[New]**: Internally routhing is using a new `MessageEnvelope`. This allows for e.g. hooking in state on a certain message being routed.